add: split the vfs abstract class implementation
